requirements.txt
pandas
openpyxl
matplotlib
numpy
python-docx
pillow
pyinstaller